Use i64 for exponent calculation

+++ vm/src/vm.rs
@@ -445,12 +445,12 @@ impl<'a, T: PrimitiveSet> Vm<'a, T> {
         } else {
             let integer = integer >> 2;
             let mantissa = if integer % 2 == 0 { 1.0 } else { -1.0 } * (integer >> 12) as f64;
-            let exponent = ((integer >> 1) % (1 << 11)) as i32 - 1023;
+            let exponent = ((integer >> 1) % (1 << 11)) as isize - 1023;
 
             Number::from_f64(if exponent < 0 {
-                mantissa / (1 << exponent.abs()) as f64
+                mantissa / (1u64 << exponent.abs()) as f64
             } else {
-                mantissa * (1 << exponent) as f64
+                mantissa * (1u64 << exponent) as f64
             })
         }
     }
